requirements.txt
gevent==21.12.0
Flask==2.1.2
lxml==4.9.1
Pillow==9.2.0